+++ database/migrations/2022_06_04_073359_create_pengembangan_wisata_table.php
@@ -16,7 +16,7 @@ class CreatePengembanganWisataTable extends Migration
         Schema::create('pengembangan_wisata', function (Blueprint $table) {
             $table->id();
             $table->integer('id_wisata');
-            $table->float('target_dana');
+            $table->bigInteger('target_dana');
             $table->timestamps();
             $table->softDeletes();
         });
